# cmd/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "飞宏IM接口",
        "title": "飞宏IM接口",
        "contact": {
            "name": "飞宏IM技术支持",
            "url": "https://github.com/liony823/chat",
            "email": "ilovecoding@foxmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:10002",
    "basePath": "/",
    "paths": {
...
        "/statistics/group/active": {
            "post": {
                "description": "获取活跃群",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "获取活跃群",
                "operationId": "getActiveGroup",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/msg.GetActiveGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/msg.GetActiveGroupResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/statistics/user/active": {
            "post": {
                "description": "获取活跃用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "获取活跃用户",
                "operationId": "getActiveUser",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/msg.GetActiveUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/msg.GetActiveUserResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/user/get_users_online_status": {
            "post": {
                "description": "获取用户在线状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户在线状态",
                "operationId": "getUsersOnlineStatus",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/msggateway.GetUsersOnlineStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/msggateway.GetUsersOnlineStatusResp_SuccessResult"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "msg.ActiveGroup": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "group": {
                    "$ref": "#/definitions/sdkws.GroupInfo"
                }
            }
        },
        "msg.ActiveUser": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/sdkws.UserInfo"
                }
            }
        },
        "msg.GetActiveGroupReq": {
            "type": "object",
            "properties": {
                "ase": {
                    "type": "boolean"
                },
                "end": {
                    "type": "integer"
                },
                "pagination": {
                    "$ref": "#/definitions/sdkws.RequestPagination"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "msg.GetActiveGroupResp": {
            "type": "object",
            "properties": {
                "dateCount": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "groupCount": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/msg.ActiveGroup"
                    }
                },
                "msgCount": {
                    "type": "integer"
                }
            }
        },
        "msg.GetActiveUserReq": {
            "type": "object",
            "properties": {
                "ase": {
                    "type": "boolean"
                },
                "end": {
                    "type": "integer"
                },
                "group": {
                    "type": "boolean"
                },
                "pagination": {
                    "$ref": "#/definitions/sdkws.RequestPagination"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "msg.GetActiveUserResp": {
            "type": "object",
            "properties": {
                "dateCount": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "msgCount": {
                    "type": "integer"
                },
                "userCount": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/msg.ActiveUser"
                    }
                }
            }
        },
        "msggateway.GetUsersOnlineStatusReq": {
            "type": "object",
            "properties": {
                "userIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "msggateway.GetUsersOnlineStatusResp_SuccessDetail": {
            "type": "object",
            "properties": {
                "connID": {
                    "type": "string"
                },
                "isBackground": {
                    "type": "boolean"
                },
                "platformID": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "msggateway.GetUsersOnlineStatusResp_SuccessResult": {
            "type": "object",
            "properties": {
                "detailPlatformStatus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/msggateway.GetUsersOnlineStatusResp_SuccessDetail"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "sdkws.GroupInfo": {
            "type": "object",
            "properties": {
                "applyMemberFriend": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "creatorUserID": {
                    "type": "string"
                },
                "ex": {
                    "type": "string"
                },
                "faceURL": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "groupName": {
                    "type": "string"
                },
                "groupType": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "lookMemberInfo": {
                    "type": "integer"
                },
                "memberCount": {
                    "type": "integer"
                },
                "needVerification": {
                    "type": "integer"
                },
                "notification": {
                    "type": "string"
                },
                "notificationUpdateTime": {
                    "type": "integer"
                },
                "notificationUserID": {
                    "type": "string"
                },
                "ownerUserID": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "sdkws.RequestPagination": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "showNumber": {
                    "type": "integer"
                }
            }
        },
        "sdkws.UserInfo": {
            "type": "object",
            "properties": {
                "appMangerLevel": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "ex": {
                    "type": "string"
                },
                "faceURL": {
                    "type": "string"
                },
                "globalRecvMsgOpt": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}